fix an edge case for remapped sequences with skips
when remapping alternate sequences, we have to "fast forward" through
the cigar string describing the alignment of said alternate sequence to
the canonical reference in order to discover the alignment to said
canonical reference. there was a bug in the code that did such that
presented for reads that aligned starting precisely after a deletion (D)
or skip (N) region.